+++ src/components/real/propulsion/initialize_simple_thruster.cpp
