+++ src/components/real/propulsion/simple_thruster.cpp
