+++ src/simulation/spacecraft/sample_spacecraft/sample_components.hpp
@@ -22,7 +22,7 @@
 #include <components/ideal/initialize_force_generator.hpp>
 #include <components/ideal/initialize_torque_generator.hpp>
 #include <components/power/power_control_unit.hpp>
-#include <components/propulsion/initialize_simple_thruster.hpp>
+#include <components/real/propulsion/initialize_simple_thruster.hpp>
 #include <dynamics/dynamics.hpp>
 #include <library/math/vector.hpp>
 #include <simulation/hils/hils_port_manager.hpp>
